pyproject.toml
[tool.poetry]
name = "fastapi-onion-architecture"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Uoiferise <ws555w@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.108.0"}
sqlalchemy = "^2.0.24"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
loguru = "^0.7.2"
black = "^23.12.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
testcontainers = "^3.7.1"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
